+++ bin/custom_multiqc.py
@@ -13,7 +13,6 @@ from constant_variables import (
     BLAST_COLUMNS,
     CONSTRAIN_GENERAL_STATS_COLUMNS,
     FILES_OF_INTEREST,
-    CLUSTER_HEADERS,
     CLUSTER_PCONFIG,
 )
 from pathlib import Path
@@ -222,66 +221,6 @@ def check_file_exists(file, throw_error=True):
     return True
 
 
-def read_header_file(file_path):
-    """
-    Read a file and return its content as a list of strings.
-
-    Args:
-        file_path (str): The path to the file to be read.
-
-    Returns:
-        list: A list of strings representing the content of the file.
-    """
-    with open(file_path, "r") as file:
-        content = file.read().splitlines()
-    return content
-
-
-def concat_table_files(table_files, **kwargs):
-    """Concatenate all the cluster summary files into a single dataframe.
-
-    Args:
-        table_files (list): List of file paths to be concatenated.
-        **kwargs: Additional keyword arguments to be passed to pd.read_csv().
-
-    Returns:
-        pd.DataFrame: The concatenated dataframe.
-    """
-    try:
-        valid_dfs = [read_file_to_dataframe(file, **kwargs) for file in table_files if check_file_exists(file)]
-
-        if not valid_dfs:
-            logging.warning("Warning concatenating files: %s", table_files)
-            logging.warning("No valid files found to concatenate.")
-            return pd.DataFrame()
-
-        df = pd.concat(valid_dfs)
-        return df
-
-    except ValueError as e:
-        logging.warning("Error concatenating files: %s\n%s", table_files, e)
-        return pd.DataFrame()
-
-
-def read_in_quast(table_files):
-    """Concatenate all the cluster summary files into a single dataframe.
-
-    Args:
-        table_files (list): List of file paths to the cluster summary files.
-
-    Returns:
-        pd.DataFrame: A dataframe containing the concatenated data from all the cluster summary files.
-    """
-    df = pd.DataFrame()
-    if table_files:
-        for file in table_files:
-            if check_file_exists(file, throw_error=False):
-                with open(file, "r") as f:
-                    d = dict(line.strip().split("\t") for line in f)
-                    df = pd.concat([df, pd.DataFrame.from_dict(d, orient="index").T])
-    return df
-
-
 def get_module_selection(table_headers: Path = None) -> Dict:
     """
     Get the files of interest and the columns of interest from the table headers file
@@ -299,29 +238,6 @@ def get_module_selection(table_headers: Path = None) -> Dict:
     yaml_data = yaml.safe_load(table_headers.read_text())
 
     return yaml_data
-
-
-def write_dataframe(df, file, comment):
-    """
-    Write a pandas DataFrame to a file in TSV format.
-
-    Args:
-        df (pd.DataFrame): The DataFrame to be written.
-        file (str): The file path to write the DataFrame to.
-        comment (list): A list of strings to be written as comments at the beginning of the file.
-
-    Returns:
-        None
-    """
-    if df.empty:
-        logger.warning("The DataFrame %s is empty, nothing will be written to the file!", file)
-        return
-    df_tsv = df.to_csv(sep="\t", index=False, quoting=csv.QUOTE_NONNUMERIC)
-    with open(file, "w") as f:
-        if comment:
-            f.write("\n".join(comment))
-            f.write("\n")
-        f.write(df_tsv)
 
 
 def filter_and_rename_columns(data: pd.DataFrame, columns: List[Union[str, Dict[str, str]]]) -> pd.DataFrame:
@@ -356,115 +272,6 @@ def filter_and_rename_columns(data: pd.DataFrame, columns: List[Union[str, Dict[
         keep_columns = [col for col in keep_columns if col in data.columns]
 
     return data[keep_columns].rename(columns=rename_dict)
-
-
-def df_from_tsv(file, **kwargs):
-    """
-    Read a dataframe from a tsv file.
-
-    Args:
-        file (str): The path to the file.
-
-    Returns:
-        pd.DataFrame: The dataframe read from the file.
-    """
-    with open(file, "r") as table:
-        df = pd.read_csv(table, sep="\t", **kwargs)
-    return df
-
-
-def df_from_csv(file, **kwargs):
-    """
-    Read a dataframe from a csv file.
-
-    Args:
-        file (str): The path to the file.
-
-    Returns:
-        pd.DataFrame: The dataframe read from the file.
-    """
-    with open(file, "r") as table:
-        df = pd.read_csv(table, **kwargs)
-    return df
-
-
-def df_from_yaml(file, **kwargs):
-    """
-    Read a dataframe from a YAML file.
-
-    Args:
-        file (str): The path to the file.
-
-    Returns:
-        pd.DataFrame: The dataframe read from the file.
-    """
-    with open(file, "r") as yaml_file:
-        data = yaml.safe_load(yaml_file, **kwargs)
-        df = pd.DataFrame(data)
-    return df
-
-
-def df_from_json(file, **kwargs):
-    """
-    Read a dataframe from a JSON file.
-
-    Args:
-        file (str): The path to the file.
-
-    Returns:
-        pd.DataFrame: The dataframe read from the file.
-    """
-    with open(file, "r") as json_file:
-        try:
-            data = json.load(json_file, **kwargs)
-        except json.JSONDecodeError as e:
-            logger.warning("Error reading JSON file %s: %s", file, e)
-            return pd.DataFrame()
-
-        # Check if 'query' key exists
-        if "filename" not in data:
-            # Get the filename without path and suffix
-            filename = os.path.splitext(os.path.basename(file))[0]
-            # Add new key-value pair
-            data["filename"] = filename + "_constrain"
-        df = pd.DataFrame([data])
-    return df
-
-
-def read_file_to_dataframe(file, **kwargs):
-    """
-    Read a dataframe from a file.
-
-    Args:
-        file (str): The path to the file.
-
-    Returns:
-        pd.DataFrame: The dataframe read from the file.
-    """
-    file_path = Path(file)
-    if os.path.getsize(file_path) == 0:
-        logger.debug("File is empty %s", file_path)
-        return pd.DataFrame()
-    if file_path.suffix in [
-        ".tsv",
-        ".txt",
-    ]:  # mqc calls tsv's txts, bed files are gzipped
-        df = df_from_tsv(file_path, **kwargs)
-    elif file_path.suffix == ".csv":
-        df = df_from_csv(file_path, **kwargs)
-    elif file_path.suffix in [".yaml", ".yml"]:
-        df = df_from_yaml(file_path, **kwargs)
-    elif file_path.suffix in [".json"]:
-        df = df_from_json(file_path, **kwargs)
-    else:
-        logger.error(
-            "The file format %s is not supported of file %s!",
-            file_path.suffix,
-            file_path,
-        )
-        sys.exit(2)
-    return df
-
 
 def join_dataframe(base_df: pd.DataFrame, dfs: List[pd.DataFrame]) -> pd.DataFrame:
     """
@@ -586,25 +393,6 @@ def read_data(directory, file_columns, process_dataframe):
         multiqc_samples_df = join_dataframe(multiqc_samples_df, [df])
 
     return multiqc_samples_df
-
-
-def get_header(comment_dir, header_file_name):
-    """
-    Get the header from a header file located in the specified comment directory.
-
-    Args:
-        comment_dir (str): The directory where the header file is located.
-        header_file_name (str): The name of the header file.
-
-    Returns:
-        list: The header as a list of strings.
-    """
-    header = []
-    if comment_dir:
-        header_file_path = f"{comment_dir}/{header_file_name}"
-        if check_file_exists(header_file_path):
-            header = read_header_file(header_file_path)
-    return header
 
 
 def dynamic_split(index_str):
@@ -937,59 +725,6 @@ def coalesce_constrain(dataframe):
 
 def fill_group_na(s):
     return s.infer_objects(copy=False).ffill().bfill()
-
-
-def read_bed(bed_files, selection):
-    """
-    Read bed files and concatenate them into a single dataframe.
-
-    Args:
-        bed_files (list): List of bed file paths.
-        selection (list): List of approved bed_files.
-
-    Returns:
-        dict: {index: pd.Dataframe} dictionary of bed dataframes.
-    """
-    bed_data = {}
-    for sample in selection:
-        bed_sel = [bed_file for bed_file in bed_files if sample_in_file(sample, bed_file)]
-        logger.debug("Bed files for %s: %s", sample, bed_sel)
-        if not bed_sel:
-            logger.warning("No bed files were found for %s", sample)
-            continue
-        bed_file = bed_sel[0]
-        if check_file_exists(bed_file):
-            logger.debug("Converting bed file %s to coverage", bed_file)
-            df = pd.read_csv(
-                bed_file,
-                sep="\t",
-                compression="gzip",
-                names=["chromosome", "position", "end", "coverage"],
-            )
-            bed_data[sample] = df[["position", "coverage"]].copy()
-
-    if len(bed_data) != len(selection):
-        logger.warning("Not all bed files were read!")
-        logger.warning("Missing bed files: %s", set(selection) - set(bed_data.keys()))
-    return bed_data
-
-
-def sample_in_file(sample, file_name) -> bool:
-    """
-    Check if the sample name is in the file name.
-
-    Args:
-        sample (str): The sample name.
-        file (str): The file name.
-
-    Returns:
-        bool: True if the sample name is in the file name, False otherwise.
-    """
-    sample_upper = sample.upper()
-    file_upper = str(file_name).upper()
-    logger.debug("sample: %s - file: %s", sample_upper, file_upper)
-
-    return sample_upper in file_upper or sample_upper in file_upper.replace("_", "-") or sample_upper in file_upper.replace("-", "_")
 
 
 def remove_keys(d, keys):
